Add modules for each standard
