+++ programs/marketplace/src/instructions/legacy/buy.rs
@@ -14,7 +14,7 @@ use tensor_toolbox::{
 use vipers::Validate;
 
 use crate::{
-    program::MarketplaceProgram, record_event, AuthorizationDataLocal, ListState, TakeEvent,
+    noop::record_event, program::MarketplaceProgram, AuthorizationDataLocal, ListState, TakeEvent,
     Target, TcompError, TcompEvent, TcompSigner, CURRENT_TCOMP_VERSION, MAKER_BROKER_PCT,
     TCOMP_FEE_BPS,
 };
