+++ programs/marketplace/src/instructions/mod.rs
@@ -1,48 +1,28 @@
-pub mod legacy;
-pub use legacy::*;
-
-pub mod token22;
-pub use token22::*;
-
-pub mod wns;
-pub use wns::*;
-
 pub mod bid;
-pub mod buy;
-pub mod buy_core;
-pub mod buy_spl;
 pub mod cancel_bid;
 pub mod close_expired_bid;
-pub mod close_expired_listing;
-pub mod close_expired_listing_core;
-pub mod delist;
-pub mod delist_core;
+pub mod compressed;
 pub mod edit;
-pub mod list;
-pub mod list_core;
+pub mod legacy;
+pub mod mpl_core;
 pub mod noop;
-pub mod take_bid_common;
-pub mod take_bid_compressed;
-pub mod take_bid_core;
+pub mod take_bid;
+pub mod token22;
 pub mod withdraw_fees;
+pub mod wns;
 
+pub use self::mpl_core::*;
 pub use bid::*;
-pub use buy::*;
-pub use buy_core::*;
-pub use buy_spl::*;
 pub use cancel_bid::*;
 pub use close_expired_bid::*;
-pub use close_expired_listing::*;
-pub use close_expired_listing_core::*;
-pub use delist::*;
-pub use delist_core::*;
+pub use compressed::*;
 pub use edit::*;
-pub use list::*;
-pub use list_core::*;
+pub use legacy::*;
 pub use noop::*;
-pub use take_bid_compressed::*;
-pub use take_bid_core::*;
+pub use take_bid::*;
+pub use token22::*;
 pub use withdraw_fees::*;
+pub use wns::*;
 
 use anchor_lang::solana_program::pubkey::Pubkey;
