+++ programs/marketplace/src/instructions/token22/close_expired_listing.rs
@@ -4,7 +4,7 @@ use anchor_spl::{
 };
 
 use self::program::MarketplaceProgram;
-use crate::*;
+use crate::{noop::record_event, *};
 
 #[derive(Accounts)]
 pub struct CloseExpiredListingT22<'info> {
