+++ programs/marketplace/src/instructions/token22/list.rs
@@ -6,8 +6,8 @@ use anchor_spl::{
 };
 
 use crate::{
-    maker_broker_is_whitelisted, program::MarketplaceProgram, record_event, ListState, MakeEvent,
-    Target, TcompError, TcompEvent, TcompSigner, CURRENT_TCOMP_VERSION, LIST_STATE_SIZE,
+    maker_broker_is_whitelisted, noop::record_event, program::MarketplaceProgram, ListState,
+    MakeEvent, Target, TcompError, TcompEvent, TcompSigner, CURRENT_TCOMP_VERSION, LIST_STATE_SIZE,
     MAX_EXPIRY_SEC,
 };
