+++ programs/marketplace/src/instructions/token22/take_bid.rs
@@ -12,7 +12,7 @@ use tensorswap::program::EscrowProgram;
 use vipers::Validate;
 
 use crate::{
-    take_bid_common::{assert_decode_mint_proof, take_bid_shared, TakeBidArgs},
+    take_bid::{assert_decode_mint_proof, take_bid_shared, TakeBidArgs},
     BidState, Field, Target, TcompError, CURRENT_TCOMP_VERSION,
 };
