+++ programs/marketplace/src/utils/shared.rs
